3-16

+++ body/src/main/java/com/adc/da/article/controller/LabelEOController.java
@@ -5,6 +5,7 @@ import static org.springframework.http.MediaType.APPLICATION_JSON_UTF8_VALUE;
 import java.util.List;
 import java.util.Map;
 
+import com.adc.da.util.utils.UUID;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -32,41 +33,20 @@ public class LabelEOController extends BaseController<LabelEO>{
     @Autowired
     private LabelEOService labelEOService;
 
-	@ApiOperation(value = "|LabelEO|分页查询")
-    @GetMapping("/page")
-    @RequiresPermissions("article:label:page")
-    public ResponseMessage<PageInfo<LabelEO>> page(LabelEOPage page) throws Exception {
-        List<LabelEO> rows = labelEOService.queryByPage(page);
-        return Result.success(getPageInfo(page.getPager(), rows));
-    }
-
 	@ApiOperation(value = "|LabelEO|查询")
-    @GetMapping("")
+    @GetMapping("/list")
     @RequiresPermissions("article:label:list")
     public ResponseMessage<List<LabelEO>> list(LabelEOPage page) throws Exception {
         return Result.success(labelEOService.queryByList(page));
 	}
 
-    @ApiOperation(value = "|LabelEO|详情")
-    @GetMapping("/{id}")
-    @RequiresPermissions("article:label:get")
-    public ResponseMessage<LabelEO> find(@PathVariable String id) throws Exception {
-        return Result.success(labelEOService.selectByPrimaryKey(id));
-    }
-
     @ApiOperation(value = "|LabelEO|新增")
-    @PostMapping(consumes = APPLICATION_JSON_UTF8_VALUE)
+    @PostMapping("/add")
     @RequiresPermissions("article:label:save")
     public ResponseMessage<LabelEO> create(@RequestBody LabelEO labelEO) throws Exception {
+	    labelEO.setId(UUID.randomUUID());
+	    labelEO.setNum(0);
         labelEOService.insertSelective(labelEO);
-        return Result.success(labelEO);
-    }
-
-    @ApiOperation(value = "|LabelEO|修改")
-    @PutMapping(consumes = APPLICATION_JSON_UTF8_VALUE)
-    @RequiresPermissions("article:label:update")
-    public ResponseMessage<LabelEO> update(@RequestBody LabelEO labelEO) throws Exception {
-        labelEOService.updateByPrimaryKeySelective(labelEO);
         return Result.success(labelEO);
     }
 
@@ -74,9 +54,13 @@ public class LabelEOController extends BaseController<LabelEO>{
     @DeleteMapping("/{id}")
     @RequiresPermissions("article:label:delete")
     public ResponseMessage delete(@PathVariable String id) throws Exception {
-        labelEOService.deleteByPrimaryKey(id);
-        logger.info("delete from label where id = {}", id);
-        return Result.success();
-    }
 
+        LabelEO labelEO = labelEOService.selectByPrimaryKey(id);
+        if (labelEO.getNum() > 0) {
+            return Result.error();
+        } else {
+            labelEOService.deleteByPrimaryKey(id);
+            return Result.success();
+        }
+    }
 }
